feat(alert): code improvements
